Break out query building
